requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openstacksdk==3.3.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
